💪 add lib-resilient
This adds lib-resilient to this website. (https://resilient.is/) It will
install a service-worker into the browsers of users. Should a user have
a service-worker installed but guardianproject.info is not accessible to
them. The lib-resilient will attempt to fetch the latest version of the
site from guardianproject.gitlab.io/info instead. This is configured in
the `config.json` file which is required by lib-resilient.