gitlab-ci: always use HTTPS for apt-get connections
The official Debian docker images ship without ca-certificates. TLS
certificates cannot be verified until that is installed. The
following includes code that turns off TLS verification, and enables
HTTPS, so at least unverified TLS is used for apt-get instead of
plain HTTP.  Once ca-certificates is installed, the CA verification
is re-enabled by removing that config.  This set up makes the initial
`apt-get update` and `apt-get install` look the same as verified TLS
to the network observer and hides the metadata.